move _search/_batch_search/search to ModelIndex
